Update patch set 1
Patch Set 1:

This is wrong. Adaptive brightness is completely unusable with these values.

The sensor can be seen with the screen off under a very bright light, it's centered with the camera notch if not slightly offset to the right.

I get a similar result with your test app in scan mode, but moving the circle around manually I've found the highest readings are with the dot at (540,70).

After rebuilding the correction service with this value, adaptive brightness is stable (doesn't fling back and forth between min and max brightness every time I change screens) and takes sensible decisions according to the environment.

I need to set myself up to submit patches here again, first time building LOS in a couple years.

Patch-set: 1
CC: Gerrit User 13106 <13106@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>